requirements.txt
PySide6
pyxtream
python-vlc
qtawesome
nuitka
